backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.1
pydantic-settings==2.2.1
google-generativeai==0.5.4
google-cloud-vision==3.7.2
httpx==0.27.0
python-dotenv==1.0.1
PyMuPDF==1.24.3
pytesseract==0.3.10
Pillow==10.3.0
aiofiles==23.2.1
pytest==8.2.0
pytest-asyncio==0.23.6